sip: fix generation of local/remote sip URIs
Various APIs return a SIP URI.
But the format is not constant: some of them
return SIP compliants, some not (ringdht), sometime
with a IP not used or not IPv6 compliants, ...
These problems causes disfonctionnalities in standard
compliants libraries like PJSIP.

This patch makes an overall pass to fix that and
ensure we use the RFC-3986 for general URI standard,
RFC-6874 for IPv6 support and RFC-3261 for SIP specifics.

The "ring.dht" hostname is now replaced by the IP
used by the underlaying SIP transport.

Change-Id: Ic612d0f742517c2c7176499281bcfd0cd68d0137